pyproject.toml
[project]
name = "cobbleseer-mcp-server"
version = "1.0.0"
description = "AI-powered Cobblemon resource pack generator - MCP Server"
authors = [
    {name = "JX_YL", email = ""}
]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
keywords = ["cobblemon", "minecraft", "mcp", "ai", "generator"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastapi>=0.104.0",
    "fastmcp>=0.5.0",
    "uvicorn[standard]>=0.24.0",
    "anthropic>=0.18.0",
    "ollama>=0.1.0",
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.0.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "jsonschema>=4.20.0",
    "loguru>=0.7.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/your-username/cobbleseer"
Documentation = "https://github.com/your-username/cobbleseer/wiki"
Repository = "https://github.com/your-username/cobbleseer"

[project.scripts]
cobbleseer = "server:main"

[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["services", "models", "utils"]

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

